Add files via upload
This simulator allows rearranging a simplified 7 card deck and changing player votes with automatic recalculation. It's meant to give the feel of the type of puzzle that setting up one player for a great score even when everyone acts completely in unison. In the real game not only will there be many more cards to deal with, you can't control the order so the best you can do is influence the odds that you'll be in a position to benefit from Eat The Rich. I'm almost positive I found the maximum single player score for this little set, it wasn't trivial.